Tweaked one of the unit tests that was failing due to rounding differences between reference file and generated file.
The test was simply using `filecmp.cmp()` to compare plain text files, which
does not cut it if the plain text files contain very precise floating figures
that may change between different platforms.

The fix consisted of comparing the text files in a bit more granular manner
(word by word), and relax the comparison if the word is parseable as a number.