TP 394631 - Removing specific prman attribute conventions when default renderer isnt prman (D3588)
Summary: Moved prmanstatements to katana attribute conventions where available and created new conventions where not.  If prman is the default renderer, we will still follow the precedent set by pixar and use prman attributes.  Also includes a fix for library naming differences between Windows and Linux.  We also remove the Pxr prefix on the node names.

Test Plan: Check that no prmanstatments appear when we are using a non-prman DEFAULT_RENDERER, and also make sure they do appear when it is set to prman

Reviewers: mark.final, #team_medusa, lea

Reviewed By: mark.final, #team_medusa, lea

Subscribers: lea

Differential Revision: https://phabricator.thefoundry.co.uk/D3588