TP 417010 - Adding Changelist and updating build tools (D4116)
Summary: Added Changelist, and added a findUSD.cmake file to make finding the USD libraries simpler.  Retained the use of pxrConfig.cmake if needed, but added the findUSD.cmake as an option.  This is especially useful when wanting to build against Katana's FnUSD.

Test Plan: Ensure that it builds and works on both Linux and windows built against Katana's USD and an external USD.

Reviewers: mark.final, #team_medusa, robert.fanner

Reviewed By: mark.final

Differential Revision: https://phabricator.thefoundry.co.uk/D4116