TP 427390 - Fixed incorrect lib prefix during USD linking using the USE_KATANA_THIRDPARTY_LIBS option (D4556)
Summary: Before we released 3.5 we changed the linux libraries for USD to use libFn prefix to fit closer to the default Linux shared library prefix and naming

Test Plan:
Try building against 3.5 and 3.6 using the USE_KATANA_THIRDPARTY_LIBS with this as an external repo.

This option is for Linux only.

Reviewers: #team_medusa, christian.aguilera

Reviewed By: #team_medusa, christian.aguilera

Subscribers: christian.aguilera

Differential Revision: https://phabricator.thefoundry.co.uk/D4556