TP 449449 - Preserve material interfaces even if there are no edits to the interface paramter values
Summary:
Ensure that material interfaces are supported and written, even if the value has not changed from the initial value of the shader.
Also makes sure that the value assigned is the same as the shader value, or it is set to the default value of the shader, if the shader is also default.