TP 463829 - [USD I/O] Read missing UsdLightShader parameters.
Summary:
This revision is the first half of D6021, which has been split in half.  This half
concerns the reading of UsdLuxLight attributes that were not previously being read.

In some cases I've preserved attributes that were read with the wrong/an old name,
for backwards compatibility.  These cases are commented.  I've also not modified
the contents of `geometry.light` since it's not clear who is using that.