TP 482198 - Converting root usd path to katana path when using IsolatePath failed
Summary: The LightList Linking logic passes through the rootPath to the ConvertUsdPathToKatLocation method.  The current implementation requires the path to be beyond the IsolatePath. However in the case of trying to find the root path, it still makes sense to return the root of where we're importing the Katana data.