TP 485194 - [USD I/O] Add a (hopefully temporary) workaround for a crash in HdStorm when referencing missing UDIM textures.
Summary:
This revision adds a workaround for a crash in HdStorm.  When loading a USD file that references UDIM
textures, Storm will attempt to bind these textures to the GPU even if the texture could not be
loaded.  This causes it to dereference a `nullptr` trying to get the `GLuint` texture handle.

In particular, trying to load a USD file containing relative UDIM textures will cause this crash, as
**UsdIn** does not yet support resolution of relative UDIMs.  The market scene makes heavy use of
relative UDIM paths.