TP 484494 - Export light and shadow linking with UsdLux Lights
Summary: This revision adds light linking support to the UsdLuxLight export. The primary difficulty here was getting the light list for each variant, this is done inside the lightLinking.py file. The modifications required for this included storing the name of the node we are baking from and using that to cook the list from each variant. This information is then passed through to the writing functions. I'm not a huge fan of passing this as "arbitraryData" but I couldn't see a nice way of getting this data to where it needed to be. Ideally I think the writing for different components of a Usd file should be contained within a class rather than class methods, allowing for some nicer access to the data required to write the file.