TP 487630 - Adding a plug-in system for UsdExport
Summary:
A plugin system has been added to UsdExport to allow external clients to write any extra data we are unable to determine, such as for custom schemas, to the USD prims on export.

They may register, de-register and inspect the plug-in registry with the API provided.  With the additional logic added to Katana, these plugins will be loaded at start-up from a UsdExportPlugins folder under the KATANA_RESOURCES paths.

This will allow us to also move some more custom, and less generic work, out of our UsdExport logic and into our own plugins which we can ship as default with KatanaUsdPlugins.  Some ideas for this already are:
- the SdrRegistry light attribute work. TP ID:487862
- our custom Schema for writing ID attributes (such that we can flawlessly roundtrip lights) TP ID:487875