TP 488816 - Handle lights with no attribute modifications or a shader assigned
Summary:
This revision adds handling for lights with empty LightParams or a non-existant material attribute through having no shader assigned. A light location is now compeltely skipped if no material attributes exist on it. If the shader name is assigned but no modifications made then an empty group attribute is created for it so the prim resolution can continue but no attributes will actually be written.