TP 491565 - Read and write 'Light' prim types to/from Katana and USD Light lists
Summary: When writing the light list, we will now traverse the whole stage looking for any locations which are "Light" prim types along with the Light-derived prim types or Light Filters. When reading a LightList, we will now add LightList tagets with the prim type "Light" to the light list.