TP 499422 - Convert from name to implementation name of shader input when importing lights and vice versa for export
Summary: We have had this problem for a long time of `lightColor` vs `color`, or `areaNormalize` vs `normalize` which is causing issues when exporting/importing lights with Prman. With this change when importing a light we convert `color` (from the USD file) to `lightColor` (to be used as Katana attribute) using the `GetImplementationName()` method. At export time we have to check if the attribute we're exporting matches the implementation name and convert (from `lightColor` to `color` in our example) it if it does.