TP 505055 - Rename folder and file names to remove Pxr prefix
Summary:
Rename folder and file names to remove Pxr prefix to avoid clashes with
op and library names with Prman shipped modules.
First step to make the code review understandable. Will not build as is.

Files have only be renamed in this review. Content of files have not been changed, this happens in a future commit to ensure clean git history.