TP 507658 - Import USD lights based on applied schema
Summary: Can now register USD import Ops against schemas. If a prim type does not have an import Op registered against it, the applied API schemas will be checked to see if there is one that has an Op registered against one of those. Currently the only schema registered is UsdLuxLightAPI. This means that lights such as PxrAovLight which have the schema applied but don't have an equivalent UsdLux light are imported automatically without having to register each type explicitly. We only expect one of the applied schemas of any prim to be registered and will warn if we encounter more.