TP 534163 - UsdIn: check isolatePath is a prefix of light paths
Summary:
The light list cache (which is just an attribute authored on a prim) is used
to populate the light list. This means that when we only import part of a
stage such as using the isolatePath paramter in the UsdIn node, the
light path might not actually exist on the stage.

This revision adds a check that the isolatePath is a prefix of the light
paths before adding them to the light list.

Also corrected a check that the prim is valid in the light list function as
UsdToKatan uses doesn't use isolatePath but provides the flattened
layer of an already masked stage to the UsdIn op.