TP-539855 - Instance source child naming could change on flushing caches
Summary: USD doesn't seem to define the order in which the `/__Prototype` Prims are created.  In the attached scene, it is possible for it to swap `/__Prototype_1` and `/__Prototype_2`.  The original method of using a set meant that this was always consistently stored in string order.  However, that meant that in fact the underlying prim data would change in ordering as the mapping of what `/Prototype_1` referenced would change.  Now using an `vector` means that we simply store it as we come across them in the USD stage traversal, which is consistent.